+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.45.8
-- **[fix]** - `NavigationMenu` autocomplete custom style border hover
+## v1.46.0
+- **[feat]** - Add new border radius `borderRadiusS` and `borderRadiusL` to theme
+- **[feat]** - Add `Logo` variant :  `ops`, `supplier` and `pricing`
+- **[feat]** - Add `Button` variant `radio`
+- **[refactor]** - Update `Button` size
+- **[refactor]** - Update `ToggleButton` style
+- **[refactor]** - Update palette colors 
+- **[refactor]** - Update `Badge` component style
+- **[refactor]** - Update `StatusIcon` component style
+- **[refactor]** - Update `Menu` component style and set min width to 200px
+- **[refactor]** - Update `Menu` & `Autocomplete` top spacing
+- **[refactor]** - Update `Snackbar` component style & default anchor origin